+++ workloads/JOB-light-Queries/8.sql
@@ -1,5 +1,0 @@
-SELECT COUNT(*)
-FROM title AS t, movie_info AS mi, movie_keyword AS mk
-WHERE t.id = mi.movie_id
-  AND t.id = mk.movie_id
-  AND t.production_year > 1990;