+++ workloads/Stats-CEB/queries/q-1.sql
@@ -1,4 +1,0 @@
-SELECT COUNT(*)
-FROM badges AS b, users AS u
-WHERE b.UserId = u.Id
-  AND u.UpVotes >= 0;