+++ workloads/Stats-CEB/queries/q-3.sql
@@ -1,5 +1,0 @@
-SELECT COUNT(*)
-FROM comments AS c, postHistory AS ph
-WHERE c.UserId = ph.UserId
-  AND c.Score = 0
-  AND ph.PostHistoryTypeId = 1;