+++ workloads/Stats-CEB/queries/q-5.sql
@@ -1,4 +1,0 @@
-SELECT COUNT(*)
-FROM comments AS c, votes AS v
-WHERE c.UserId = v.UserId
-  AND c.Score = 0;